+++ src/indigo/interfaces/logic/enums/PREDICATE_TYPE.java
@@ -1,5 +1,5 @@
 package indigo.interfaces.logic.enums;
 
 public enum PREDICATE_TYPE {
-	bool, numeric, _
+	bool, Int, _
 }
